renamed and moved tasks (WIP)
