+++ backend/onyx/background/celery/versioned_apps/docfetching.py
@@ -10,7 +10,7 @@ set_is_ee_based_on_env_variable()
 
 
 def get_app() -> Celery:
-    from onyx.background.celery.apps.docprocessing import celery_app
+    from onyx.background.celery.apps.docfetching import celery_app
 
     return celery_app
