+++ backend/onyx/configs/app_configs.py
@@ -327,18 +327,18 @@ except ValueError:
 CELERY_WORKER_KG_PROCESSING_CONCURRENCY = int(
     os.environ.get("CELERY_WORKER_KG_PROCESSING_CONCURRENCY") or 4
 )
-CELERY_WORKER_DOCPROCESSING_CONCURRENCY_DEFAULT = 1
+CELERY_WORKER_DOCFETCHING_CONCURRENCY_DEFAULT = 1
 try:
-    env_value = os.environ.get("CELERY_WORKER_DOCPROCESSING_CONCURRENCY")
+    env_value = os.environ.get("CELERY_WORKER_DOCFETCHING_CONCURRENCY")
     if not env_value:
-        env_value = os.environ.get("NUM_DOCPROCESSING_WORKERS")
+        env_value = os.environ.get("NUM_DOCFETCHING_WORKERS")
 
     if not env_value:
-        env_value = str(CELERY_WORKER_DOCPROCESSING_CONCURRENCY_DEFAULT)
-    CELERY_WORKER_DOCPROCESSING_CONCURRENCY = int(env_value)
+        env_value = str(CELERY_WORKER_DOCFETCHING_CONCURRENCY_DEFAULT)
+    CELERY_WORKER_DOCFETCHING_CONCURRENCY = int(env_value)
 except ValueError:
-    CELERY_WORKER_DOCPROCESSING_CONCURRENCY = (
-        CELERY_WORKER_DOCPROCESSING_CONCURRENCY_DEFAULT
+    CELERY_WORKER_DOCFETCHING_CONCURRENCY = (
+        CELERY_WORKER_DOCFETCHING_CONCURRENCY_DEFAULT
     )
 
 # The maximum number of tasks that can be queued up to sync to Vespa in a single pass
