+++ backend/supervisord.conf
@@ -89,12 +89,12 @@ autorestart=true
 startsecs=10
 stopasgroup=true
 
-[program:celery_worker_docprocessing]
-command=celery -A onyx.background.celery.versioned_apps.docprocessing worker
+[program:celery_worker_docfetching]
+command=celery -A onyx.background.celery.versioned_apps.docfetching worker
     --loglevel=INFO
-    --hostname=docprocessing@%%n
+    --hostname=docfetching@%%n
     -Q connector_document_extraction_task
-stdout_logfile=/var/log/celery_worker_docprocessing.log
+stdout_logfile=/var/log/celery_worker_docfetching.log
 stdout_logfile_maxbytes=16MB
 redirect_stderr=true
 autorestart=true
@@ -173,7 +173,7 @@ command=tail -qF
     /var/log/celery_worker_heavy.log
     /var/log/celery_worker_indexing.log
     /var/log/celery_worker_user_files_indexing.log
-    /var/log/celery_worker_docprocessing.log
+    /var/log/celery_worker_docfetching.log
     /var/log/celery_worker_monitoring.log
     /var/log/slack_bot.log
     /var/log/supervisord_watchdog_celery_beat.log
